+++ components/BUILD.gn
@@ -43,6 +43,7 @@ test("brave_components_unittests") {
     "//brave/components/json:unit_tests",
     "//brave/components/local_ai/core:unit_tests",
     "//brave/components/oblivious_http:unit_tests",
+    "//brave/components/screenshot/core/browser:unit_tests",
     "//brave/components/static_redirect_helper:unit_tests",
     "//brave/components/time_period_storage:unit_tests",
     "//brave/components/update_client:unit_tests",
