+++ components/screenshot/core/browser/DEPS
@@ -1,3 +1,4 @@
 include_rules = [
   "+third_party/skia/include/core",
+  "+ui/gfx/image",
 ]
